requirements.txt
cassio 
langchain-community
openai 
tiktoken
PyPDF2
streamlit
langchain_openai